DEV-5653: split up views for switchboard and revengine
